break individual parts of model out into functions, and allow both analyses from the same class
